% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_markov.r
\name{copula_markov}
\alias{copula_markov}
\title{Copula Markov Model}
\usage{
copula_markov(margin, copula)
}
\arguments{
\item{margin}{Package-defined margin class.}

\item{copula}{Copula object from the \code{copula} package.}
}
\value{
A \code{copula_markov} object.
}
\description{
Construct a copula Markov model from its margin and copula
}
