% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaMarkov-package.R
\docType{package}
\name{CopulaMarkov-package}
\alias{CopulaMarkov}
\alias{CopulaMarkov-package}
\title{CopulaMarkov: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrea Ferrero \email{andrea.ferrero@umontpellier.fr}

Authors:
\itemize{
  \item Andrea Ferrero \email{andrea.ferrero@umontpellier.fr}
}

}
\keyword{internal}
