% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r, R/margin_egpd.r
\name{d_margin}
\alias{d_margin}
\alias{d_margin.egpd}
\title{Density of a margin}
\usage{
d_margin(margin, x, log = FALSE, ...)

\method{d_margin}{egpd}(margin, x, log = FALSE, ...)
}
\arguments{
\item{margin}{A \code{margin} object.}

\item{x}{Numeric vector of quantiles.}

\item{log}{Logical to indicate whether to return the log density}

\item{...}{Passed to methods.}
}
\value{
Numeric vector of log-density values.
}
\description{
Density of a margin
}
\section{Methods (by class)}{
\itemize{
\item \code{d_margin(egpd)}: EGPD margin, via \code{egpd::degpd_density()}.

}}
