% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_egpd.r
\name{egpd_margin}
\alias{egpd_margin}
\title{Extended Generalized Pareto margin}
\usage{
egpd_margin(par = NULL)
}
\arguments{
\item{par}{Named parameter vector (kappa, sigma, xi)}
}
\value{
An object of class \code{c("egpd", "margin")}.
}
\description{
Constructs an EGPD margin with parameters \code{kappa}, \code{sigma} and \code{xi},
backed by the \pkg{egpd} package.
}
