% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_markov.r
\name{fit}
\alias{fit}
\title{Fit a Copula Markov Model}
\usage{
fit(object, init, x, ...)
}
\arguments{
\item{object}{A Copula Markov Model object.}

\item{init}{Initial parameter vector on the natural parameter scale.}

\item{x}{Observed data.}

\item{...}{Passed to methods.}
}
\value{
Fitted parameter vector
}
\description{
Fit a Copula Markov Model
}
