% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_markov.r
\name{log_lik.copula_markov}
\alias{log_lik.copula_markov}
\title{Log-likelihood of a Copula Markov Model}
\usage{
\method{log_lik}{copula_markov}(object, par, x, ...)
}
\arguments{
\item{object}{Copula Markov Model object.}

\item{par}{Parameter vector on the natural parameter scale.}

\item{x}{Data vector.}

\item{...}{Passed to methods.}
}
\value{
Numeric log-likelihood.
}
\description{
Log-likelihood of a Copula Markov Model
}
