% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_markov.r
\name{log_lik}
\alias{log_lik}
\title{Log-likelihood of a model}
\usage{
log_lik(object, par, x, ...)
}
\arguments{
\item{object}{A model object.}

\item{par}{Parameter vector on the natural parameter scale.}

\item{x}{Observed data.}

\item{...}{Passed to methods.}
}
\value{
Numeric log-likelihood.
}
\description{
Log-likelihood of a model
}
