% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r, R/margin_egpd.r
\name{p_margin}
\alias{p_margin}
\alias{p_margin.egpd}
\title{Cumulative distribution function of a margin}
\usage{
p_margin(margin, x, ...)

\method{p_margin}{egpd}(margin, x, ...)
}
\arguments{
\item{margin}{A \code{margin} object.}

\item{x}{Numeric vector of quantiles.}

\item{...}{Passed to methods.}
}
\value{
Numeric vector of probabilities in (0, 1).
}
\description{
Cumulative distribution function of a margin
}
\section{Methods (by class)}{
\itemize{
\item \code{p_margin(egpd)}: EGPD margin, via \code{egpd::pegpd()}.

}}
