% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r
\name{param_update}
\alias{param_update}
\title{Utility to update margin parameters}
\usage{
param_update(margin, new_param)
}
\arguments{
\item{margin}{A margin object}

\item{new_param}{Vector of new parameter values}
}
\description{
Utility to update margin parameters
}
