% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r
\name{print.margin}
\alias{print.margin}
\title{Print a margin}
\usage{
\method{print}{margin}(x, ...)
}
\arguments{
\item{x}{A \code{margin} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a margin
}
