% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r, R/margin_egpd.r
\name{q_margin}
\alias{q_margin}
\alias{q_margin.egpd}
\title{Quantile function of a margin}
\usage{
q_margin(margin, p, ...)

\method{q_margin}{egpd}(margin, p, ...)
}
\arguments{
\item{margin}{A \code{margin} object.}

\item{p}{Numeric vector of probabilities in (0, 1).}

\item{...}{Passed to methods.}
}
\value{
Numeric vector of quantiles.
}
\description{
Quantile function of a margin
}
\section{Methods (by class)}{
\itemize{
\item \code{q_margin(egpd)}: EGPD margin, via \code{egpd::qegpd()}.

}}
