% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_class.r, R/margin_egpd.r
\name{r_margin}
\alias{r_margin}
\alias{r_margin.egpd}
\title{Random generation from a margin}
\usage{
r_margin(margin, n, ...)

\method{r_margin}{egpd}(margin, n, ...)
}
\arguments{
\item{margin}{A \code{margin} object.}

\item{n}{Number of observations to draw.}

\item{...}{Passed to methods.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Random generation from a margin
}
\section{Methods (by class)}{
\itemize{
\item \code{r_margin(egpd)}: EGPD margin, via \code{egpd::regpd()}.

}}
