% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_markov.r
\name{simulate.copula_markov}
\alias{simulate.copula_markov}
\title{Simulate Copula Markov Model using Inverse Rosenblatt Transform}
\usage{
\method{simulate}{copula_markov}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{Copula Markov model.}

\item{nsim}{Sample size.}

\item{seed}{Optional random seed.}

\item{...}{Passed to methods.}
}
\value{
A list containing data on the margin scale (\code{X}) and copula
scale (\code{U}).
}
\description{
Simulate Copula Markov Model using Inverse Rosenblatt Transform
}
